spike: complete SDL3 platform migration + tab drag, disk scrollback, snap-resize
internal/platform/ is now the live platform layer; internal/app drives
it instead of cimgui-go/backend/sdlbackend. Plus several features.

Migration phases (from docs/SDL3_PLAN.md):
- Phase 0/1a/1b: SDL3+GL3 shell, Go-side render callback via cimgui-go,
  real terminal renderer wired into platform.Frame
- Phase 2: xdg_popup-backed context menu using SDL_CreatePopupWindow.
  Required wlgrab.c to issue the xdg_popup.grab(seat, serial) call SDL3
  silently omits, and xgrab.c for the X11 XGrabPointer equivalent.
  Popup uses impl_sdlrenderer3 in a separate ImGui context to avoid
  multi-context impl_opengl3 SIGSEGVs.
- Phase 4: refactor internal/app onto internal/platform. Deletes
  runloop.go, sdl_helpers.go, cellsnap_*.go, cellsnap_*.m. sdlhack
  ported SDL2 -> SDL3. Multi-viewport patched on in impl_sdl3 (upstream
  gates it off Wayland because of no-global-mouse).

Features built on top of the migration:
- Cell-snap resize: XSetWMNormalHints on X11, NSWindow.contentResize-
  Increments on macOS via new internal/platform/cellsnap_{linux,darwin}
  files. No-op on Wayland (no protocol).
- Unlimited scrollback (VTE-style): internal/terminal/diskscrollback.go
  spills evicted lines to a temp file (auto-unlinked at create), serves
  random-access reads back to the renderer. Triggered when Scrollback
  Mode = "unlimited" in prefs; ~4096-line live window in vt's RAM ring
  with the rest on disk.
- Tab drag-to-reorder within a window's tab bar, plus cross-window
  drag using Wayland data_device with snapshot-at-drop-event handling
  (g_drop_target captured in dd_drop before subsequent dd_leave can
  clear it; src_dnd_drop_performed / src_cancelled fall back to
  g_last_drag_target). Release-time SDL_GetMouseFocus is the
  non-Wayland fallback. App.updateTabDragDrop centralizes the drop
  resolution so per-Window iteration order doesn't race.
- Default to SDL_VIDEODRIVER=wayland over xwayland on Wayland sessions
  via SDL_HINT_VIDEO_DRIVER (xwayland fundamentally can't honor
  cross-app pointer grabs, which broke popup dismiss + tab drag).

+++ internal/sdlhack/mouse.go
@@ -1,13 +1,17 @@
-// Package sdlhack provides macOS workarounds for SDL2 input quirks that
-// cimgui-go's standard backend doesn't handle. SDL2 + Cocoa occasionally
+// Package sdlhack provides macOS workarounds for SDL input quirks that
+// cimgui-go's standard backend doesn't handle. SDL + Cocoa occasionally
 // drops mouse-up events after a window-focus shift, leaving SDL with a
 // stuck "button still held" state until the user app-switches. We bypass
 // the event queue by reading the OS-level button state directly.
+//
+// Ported to SDL3 as part of the platform migration. SDL3 changed mouse
+// coordinate types from int to float and renamed SDL_GetWindowSize is
+// still int-out, SDL_GetGlobalMouseState now writes floats.
 package sdlhack
 
 /*
-#cgo pkg-config: sdl2
-#include <SDL2/SDL.h>
+#cgo pkg-config: sdl3
+#include <SDL3/SDL.h>
 
 // Returns 1 iff the OS-level cursor position is inside the content
 // rect of whichever SDL_Window currently has mouse focus (any of the
@@ -16,15 +20,15 @@ package sdlhack
 // means AppKit consumed any click there and we shouldn't synthesize
 // a fake terminal click out of it.
 static int xerotty_mouse_in_window_content(void) {
-	int gx, gy;
+	float gx, gy;
 	SDL_GetGlobalMouseState(&gx, &gy);
 	SDL_Window *win = SDL_GetMouseFocus();
 	if (!win) return 0;
 	int wx, wy, ww, wh;
 	SDL_GetWindowPosition(win, &wx, &wy);
 	SDL_GetWindowSize(win, &ww, &wh);
-	if (gx < wx || gx >= wx + ww) return 0;
-	if (gy < wy || gy >= wy + wh) return 0;
+	if (gx < (float)wx || gx >= (float)(wx + ww)) return 0;
+	if (gy < (float)wy || gy >= (float)(wy + wh)) return 0;
 	return 1;
 }
 */
@@ -34,7 +38,7 @@ import "C"
 // state of the left mouse button. Returns true iff the button is currently
 // physically held according to the OS.
 func LeftButtonGlobalDown() bool {
-	var x, y C.int
+	var x, y C.float
 	state := C.SDL_GetGlobalMouseState(&x, &y)
 	return uint32(state)&C.SDL_BUTTON_LMASK != 0
 }
@@ -44,7 +48,7 @@ func LeftButtonGlobalDown() bool {
 // position, so it stays accurate even when SDL drops mouse events
 // during Cocoa focus shifts.
 func GlobalMousePos() (int, int) {
-	var x, y C.int
+	var x, y C.float
 	C.SDL_GetGlobalMouseState(&x, &y)
 	return int(x), int(y)
 }
